+++ qiime2/sdk/tests/data/has_nan.html
@@ -1,0 +1,10 @@
+<table>
+    <tr>
+        <th>NaN</th>
+        <th>Not NaN</th>
+    </tr>
+    <tr>
+        <td>NaN</td>
+        <td>42</td>
+    </tr>
+ </table>
