+++ qiime2/sdk/tests/data/no_nan.html
@@ -1,0 +1,10 @@
+<table>
+    <tr>
+        <th>Not NaN</th>
+        <th>Also Not NaN</th>
+    </tr>
+    <tr>
+        <td>42</td>
+        <td>43</td>
+    </tr>
+ </table>
